+++ ndk-hooks/src/mutes/store/init.ts
@@ -7,15 +7,20 @@ import type { Hexpubkey } from "@nostr-dev-kit/ndk";
  * @param get Zustand get function
  * @param pubkey The user's public key
  */
-export const initMutes = (set: (state: any) => void, get: () => NDKMutesState, pubkey: Hexpubkey) => {
-    set((state: any) => {
-        if (!state.mutes.has(pubkey)) {
-            state.mutes.set(pubkey, {
-                pubkeys: new Set<Hexpubkey>(),
-                hashtags: new Set<string>(),
-                words: new Set<string>(),
-                eventIds: new Set<string>(),
-            });
-        }
+export const initMutes = (
+    set: (partial: Partial<NDKMutesState> | ((state: NDKMutesState) => Partial<NDKMutesState>)) => void,
+    get: () => NDKMutesState,
+    pubkey: Hexpubkey
+) => {
+    set((state) => {
+        if (state.mutes.has(pubkey)) return {};
+        const newMutes = new Map(state.mutes);
+        newMutes.set(pubkey, {
+            pubkeys: new Set<Hexpubkey>(),
+            hashtags: new Set<string>(),
+            words: new Set<string>(),
+            eventIds: new Set<string>(),
+        });
+        return { mutes: newMutes };
     });
 };